Lots and lots of updates - see details below
*	Fixed a bug where initial solutions may include invalid centroid
	connectors
*	New calculator has been added to determine the lowest cost path
	costs, which are calculated in parallel
*	AEC calculation has been rewritten to depend on new calculator (this
	fixes an earlier issue where AEC didn't work at all)
*	GapCalculator now also depends on the new calculator (this is a change
	in behavior - previously, only the lowest-cost USED paths were
	considered. Old behavior is left as a comment, may be extracted later)
*	A Bush's links can now be retrieved as a Stream if needed
*	BushMerges now store double-precision splits (fixes numerical issues,
	but invalidates all previously-found solution files)
*	Added methods to Bush to calculate its total incurred cost and demand,
	as well as check the validity of used CentroidConnectors and flow
	conservation (these validity checks are currently not being called)
*	TSGCCalculator has been rewritten to use the new Bush total incurred
	cost method, which now allows this to be calculated in parallel
*	Re-enabled full network statistics output
*	Added a method to BushOptimizer that checks that all Links have the
	same amount of flow as the sum of the Bushes' flows on that link (this
	is currently not being called)
*	FibonacciHeap now uses double-precision labels
*	Added a (commented-out) mechanism to alter the number of inner
	iterations through a command-line argument
*	Added classpath and all .bush (saved structure) files to gitignore

+++ src/edu/utexas/wrap/util/FibonacciHeapTest.java
@@ -19,7 +19,7 @@ class FibonacciHeapTest {
 		random.setSeed(90210L);
 		
 		for (Integer i = 0; i < numCases; i++) {
-			Float key = random.nextFloat();
+			Double key = random.nextDouble();
 			//System.out.println(key);
 			Node p = new Node(i,false,i);
 			map.put(i, p);
